refactor: ApplyDetail 파일구조 수정(#239)
